fix: 코드설명 주석 더함
